Fix error in SimulationContext >> #runUntilErrorOrReturnFrom:
Confusingly, #caseOf:otherwise: does not check the receiver and the passed associations' keys' block values for identity but for equality. This caused the simulation loop to stop too early when certain #ensure: contexts was reached. The rest of the simulation was then run in the #stepToCallee loop where every UnhandledError would be resumed without aborting the simulation.

Closes #10.