Fixes simulation of mirror primitives in Sandbox
Instead of overwriting the mirror primitives themselves, override their relevant senders from the 'instruction decoding' protocol and forward their sends to the #simulatedObject:... alternatives that are added on SandboxContext. Additionally, in #doPrimitive:method:receiver:args:, make sure to redirect the arguments at the right indices, which are shifted by one when using the mirrored version of a primitive.

Adds another prevdep for Kernel-ct.1383 (inbox) to make all MirrorPrimitiveTests pass that are referred to in the SimulationKernelTests via 492979e.

Closes #20.