Clean up, fix, and test simulation of BitBlt primitives
Nuke the entire object preparation/update protocol and the ridiculuous hacks for BitBlt instances from Sandbox but use an access indirector in #doNamedPrimitiveIn:for:withArgs: instead. Fix unintended breakouts from the sandbox by adding manually post-copying of the destForm's bits which are accessed by the BitBltPlugin. Add a regression test for the breakout.