Fix sandboxed simulation of messages not understood by the executor
Previously, the sandbox kernel tests for ContextTest >> #testMessageNotUnderstood failed because when attempting to store an instance of TestEmptyClass in the sandbox memory, the messages #identityHash and #== were sent to this instance by WeakIdentityKeyDictionary. To resolve that error, this patch changes the dictionary type in sandboxes to a PluggableWeakKeyDictionary which uses mirror primitives for hashing and comparing the objects in the sandbox memory.